// src/app/views/biens/models/surface.ts
export interface Surface {
    surfaceRDC: number;
    surfaceRDCFonderee: number;
    surfaceVente: boolean;
    surfaceSousSol: number;
    surfaceSousSolFonderee: number;
    surface1Etage: number;
    surface1Fonderee: number;
    surfaceAutre: number;
    surfaceTotale: number;
    surfaceFonderee: number;
    nouvelleSurfaceFonderee: number;
}